splitwiser-backend/requirements.txt
# Python dependencies
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
